Move some vm tests to tests file
